Protected the API routes by making the requireAuth.js middlewear. In addition to that, I made authorized requests which will only allow logged in users to see workouts and manipulate the data.

+++ frontend/src/components/WorkoutDetails.js
@@ -1,10 +1,17 @@
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
-
+import { useAuthContext } from '../hooks/useAuthContext';
 const WorkoutDetails = ({ workout }) => {
+
+    const { user } = useAuthContext();
     // A function to delete the workout.
     const handleClick = async () => {
+
+        if(!user){
+            return;
+        }
         const response = await fetch(`/api/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {  "Authorization": `Bearer ${user.token}`}
         });
         const json = await response.json();
         if (response.ok) {
